// tests/src/bindings/CallbackCapturer.types.ts
export type Binary = string;
// ...
export type StdAck = {
  result: Binary;
} | {
  error: string;
};